Add conceptType filter Delete temp controller file and move functions.
