+++ concept-api/src/main/scala/no/ndla/conceptapi/service/search/DraftConceptSearchService.scala
@@ -128,7 +128,10 @@ trait DraftConceptSearchService {
     }
 
     def executeSearch(queryBuilder: BoolQuery, settings: DraftSearchSettings): Try[SearchResult[api.ConceptSummary]] = {
-      val idFilter      = if (settings.withIdIn.isEmpty) None else Some(idsQuery(settings.withIdIn))
+      val idFilter = if (settings.withIdIn.isEmpty) None else Some(idsQuery(settings.withIdIn))
+      val typeFilter = Option.when(settings.conceptType.nonEmpty) {
+        termsQuery("conceptType", settings.conceptType.get)
+      }
       val statusFilter  = boolStatusFilter(settings.statusFilter)
       val subjectFilter = orFilter(settings.subjects, "subjectIds")
       val tagFilter     = languageOrFilter(settings.tagsToFilterBy, "tags", settings.searchLanguage, settings.fallback)
@@ -149,6 +152,7 @@ trait DraftConceptSearchService {
       val filters =
         List(
           idFilter,
+          typeFilter,
           languageFilter,
           subjectFilter,
           tagFilter,
